Add Restart function to Framework and fix bugs in Sokoban game
In this commit, I've added a Restart function to the Framework, which will allow players to restart the game without having to reload the entire game. This should improve the user experience and make the game more convenient to play.

I've also fixed several bugs in the Sokoban game, including an issue with the Load/Save function, which was not working correctly, and a bug with the Resume Game function, which was causing a wierd GLITCH IN THE MATRIX.

Signed-off-by: Thomas <dommyhelp@gmail.com>